+++ src/port/saturn/gfx/saturn_geo_state_observer.h
@@ -1,0 +1,6 @@
+#ifndef SM64_SATURN_GEO_STATE_OBSERVER_H
+#define SM64_SATURN_GEO_STATE_OBSERVER_H
+
+#include "saturn_actor_instance.h"
+
+#endif